PD#135579: amadec: fix amstream_do_ioctl error

+++ amadec/adec-pts-mgt.h
@@ -26,7 +26,7 @@ typedef enum {
 #define AMSTREAM_IOC_GET_LAST_CHECKIN_VPTS   _IOR(AMSTREAM_IOC_MAGIC, 0xaa, unsigned long)
 #define AMSTREAM_IOC_GET_LAST_CHECKOUT_APTS  _IOR(AMSTREAM_IOC_MAGIC, 0xab, unsigned long)
 #define AMSTREAM_IOC_GET_LAST_CHECKOUT_VPTS  _IOR(AMSTREAM_IOC_MAGIC, 0xac, unsigned long)
-#define AMSTREAM_IOC_AB_STATUS  _IOR(AMSTREAM_IOC_MAGIC, 0x09, unsigned long)
+#define AMSTREAM_IOC_AB_STATUS  _IOR(AMSTREAM_IOC_MAGIC, 0x09, int)
 
 #define TSYNC_PCR_DISPOINT  "/sys/class/tsync_pcr/tsync_pcr_discontinue_point"
 #define TSYNC_PCRSCR    "/sys/class/tsync/pts_pcrscr"
